Try splitting source files
